+++ rolling/__init__.py
@@ -3,6 +3,7 @@ from .entropy import Entropy
 from .hash import PolynomialHash
 from .logical import All, Any
 from .matching import Match
+from .monotonic import Monotonic
 from .minmax import Min, Max, MinHeap
 from .similarity import JaccardIndex
 from .stats import Mean, Var, Std, Median, Mode, Skew, Kurtosis